Moving NewIndexInfo, NewRepositoryInfo and associated helpers into config.go
